+++ CHANGELOG.md
@@ -11,6 +11,7 @@ in github. Commentary on the change should appear as a nested, unordered list.
 
 - Bugfixes
   - Fix the handling of --ns-regex and --test-ns-regex (#183)
+  - Read options from the Leiningen project file (#155) 
 
 ## 1.0.10
 - Improvements